src/obj_parser/tests.rs
use super::*;

mod unit_tests {
    use super::*;

    #[test]
    fn parser_should_ignore_unrecognised_lines() {
        let invalid_obj_file = "There was a young lady named Bright
who traveled much faster than light.
She set out one day
in a relative way,
and came back the previous night.";

        let out = parse(invalid_obj_file);
        assert_eq!(out.ignored_lines, 5);
    }

    #[test]
    fn parser_should_parse_vertex_data() {
        let input = "v -1 1 0
        v -1.0000 0.5000 0.0000
        v 1 0 0
        v 1 1 0";

        let out = parse(input);
        assert_eq!(out.vertex(1), Some(Point3D::new(-1.0, 1.0, 0.0)));
        assert_eq!(out.vertex(2), Some(Point3D::new(-1.0, 0.5, 0.0)));
        assert_eq!(out.vertex(3), Some(Point3D::new(1.0, 0.0, 0.0)));
        assert_eq!(out.vertex(4), Some(Point3D::new(1.0, 1.0, 0.0)));
    }
}
